Move get_logs alongside other service-related functions in portServiceChecker.py
