+++ src/app/portServiceChecker.py
@@ -12,7 +12,7 @@
 import subprocess, re, socket, requests
 from datetime import datetime
 from time import sleep
-from app.extensions import logger
+from app.extensions import globals, logger
 
 
 def checkLocalPorts() -> str:
@@ -334,3 +334,70 @@ def checkServiceLoop(service: dict, interval: int = 30, max_checks: int = 0) -> 
             logger.info(f"Reached the maximum number of service checks ({max_checks}).. Will no longer check on service {service}.")
             return True
         sleep(interval)
+
+
+def get_logs(service: dict) -> None:
+    """Get the logs of a logged service by using SSH.
+
+    Args:
+        service (dict): Service dictionary
+    """
+
+    log_time = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
+    while True:
+        sleep(10)
+        error = ''
+        try:
+            ## `systemctl is-active` to get/log service info
+            get_status_cmd = f"sshpass -p {service['password']} ssh -o StrictHostKeyChecking=no {service['user']}@{service['host']} 'systemctl is-active {service['service']}'"
+            status_response = subprocess.run(get_status_cmd, shell=True,capture_output=True, timeout=10)
+        except subprocess.TimeoutExpired:
+            logger.error(f"SERVICE_LOGGER: request to host timed out.")
+            globals.services_status[service['service']] = [service['host'],f"request to host timed out. Trying again shortly"]
+            continue
+        except Exception as e:
+            logger.error(f"SERVICE_LOGGER: Exception attempting to retrieve service status: {e}")
+            globals.services_status[service['service']] = [service['host'],f"Exception attempting to retrieve service status."]
+            continue
+        if status_response.stdout.decode('utf-8') == '':
+            logger.error(f"SERVICE_LOGGER: Failed to get service status. Error: {status_response.stderr.decode('utf-8')}")
+            globals.services_status[service['service']] = [service['host'],f"Error has occurred when attempting to get service status."]
+            continue
+        if status_response.stdout.decode('utf-8').strip('\n') == 'failed':
+            error = "SERVICE FAILED: "
+            logger.error(f"SERVICE_LOGGER: {error} {service['service']}")
+            globals.services_status[service['service']] = [service['host'],f"Service is in failed state."]
+        elif status_response.stdout.decode('utf-8').strip('\n') == 'inactive':
+            error = "SERVICE INACTIVE: "
+            logger.error(f"SERVICE_LOGGER: {error} {service['service']}")
+            globals.services_status[service['service']] = [service['host'],f"Service is in inactive state."]
+        else:
+            globals.services_status[service['service']] = [service['host'],f"Service is in active state."]
+        ## grab logs
+        try:
+            get_log_cmd = f"sshpass -p {service['password']} ssh -o StrictHostKeyChecking=no {service['user']}@{service['host']} 'journalctl --since \"{log_time}\" -u {service['service']}'"
+            log_response = subprocess.run(get_log_cmd, shell=True,capture_output=True,timeout=10)
+            cur_logs = log_response.stdout.decode('utf-8')
+        except subprocess.TimeoutExpired:
+            logger.error(f"SERVICE_LOGGER: SSH connection to {service['user']}@{service['host']} timed out.")
+            continue
+        except Exception as e:
+            logger.error(f"SERVICE_LOGGER: Exception attempting to retrieve logs: {e}")
+            continue
+        if log_response.stdout.decode('utf-8') == '':
+            logger.error(f"SERVICE_LOGGER: Failed to collect logs. Error: {log_response.stderr.decode('utf-8')}")
+            continue
+        if "No entries" in cur_logs:
+            if error:
+                logger.error(f"SERVICE_LOGGER: {error} No new logs fe.")
+            else:
+                logger.info(f"SERVICE_LOGGER: No new logs found at this time.")
+            continue
+        output = cur_logs.split("\n")
+        output.remove("")
+        for line in output:
+            if error == False:
+                logger.info(f"SERVICE_LOGGER: {line}")
+            else:
+                logger.error("SERVICE_LOGGER: "+ error + line)
+        log_time = datetime.datetime.now().strftime('%Y-%m-%d %H:%M:%S')
